backend/requirements.txt
fastapi
uvicorn
psutil
numpy
scikit-learn
mysql-connector-python
